# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Stock Analysis Platform API
  version: 1.0.0
  description: RESTful API for stock analysis and portfolio management
  contact:
    name: API Support
    email: support@example.com
servers:
  -
    url: "http://localhost:3000"
    description: Development server
  -
    url: "https://api.example.com"
    description: Production server
paths:
...
  /api/:symbol/changes:
    get:
      summary: Get rating change trend for a stock
      description: "Implements Requirement 13.5: 评级变化追踪 - 支持查看变化趋势 (实现需求: 13.5)"
      tags:
        - Quant Rating
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: symbol
          in: path
          description: symbol parameter
          required: true
          schema:
            type: string
...
  /api/:symbol/history:
    get:
      summary: Get historical OHLCV data for a stock
      description: "Implements Requirement 4.3: Dynamically update chart to display data for selected time range Path Parameters: - symbol: Stock symbol (required) Query Parameters: - range: Time range (optional, default: '1M') Valid values: '1D', '5D', '1M', '3M', '6M', '1Y', '5Y', 'MAX' (实现需求: 4.3)"
      tags:
        - Stocks
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: symbol
          in: path
          description: Stock symbol
          required: true
          schema:
            type: string
        -
          name: range
          in: path
          description: Time range
          required: false
          schema:
            type: string
  /api/:symbol/calculate:
    post:
      summary: Calculate and save a new quant rating for a stock
      description: "This endpoint triggers a fresh calculation of the quant rating and sends notifications if the rating changes Implements Requirements: - 13.5: 评级变化追踪 - 记录评级历史 - 13.6: 评级变化推送 - 评级变化时推送通知"
      tags:
        - Quant Rating
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                      id: 1
                      createdAt: "2024-01-01T00:00:00Z"
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: symbol
          in: path
          description: symbol parameter
          required: true
          schema:
            type: string
  /api/sector/:sector/rankings:
    get:
      summary: Get sector rankings for all stocks in a sector
      description: "Implements Requirement 13.4: 显示该股票在板块和行业中的排名 (实现需求: 13.4)"
      tags:
        - Quant Rating
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: sector
          in: path
          description: sector parameter
          required: true
          schema:
            type: string
  /api/industry/:industry/rankings:
    get:
      summary: Get industry rankings for all stocks in an industry
      description: "Implements Requirement 13.4: 显示该股票在板块和行业中的排名 (实现需求: 13.4)"
      tags:
        - Quant Rating
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: industry
          in: path
          description: industry parameter
          required: true
          schema:
            type: string
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...